Cleanup part 2
